Refactor keys & add native `KeyReader`
- Refactor keys into separate elements for key modifiers and key codes

- Implement `KeyReader` on Native which causes many changes to the
  implementation:

  - Add non-blocking read

  - Switch `fgetc` to `read` so they don't use the same buffer

  - Switching modes take effect immediately and doesn't flush
    input (flushing input means we lose key codes)

  - Don't translate CR to NL

- Implement `Prompt` as a generic (cross-backend) example

For efficiency we may need to rethink how we handle input to store a
character buffer. I think the longest key code is 4 characters, so we
could read 4 characters at a time and keep a buffer of elements. Dealing
with unicode (i.e. characters longer than one byte) is also an open question.